examples/strands-agents/requirements.txt
strands-agents
strands-agents-tools
camel-ai